refactor: add-new-product 컴포넌트 분리
